Add routine to calculate an ideal swappiness value
Typically "desktop distros" are applying low swappiness by default, usually
with a value of 10, in a "one size fits all" approach. Unfortunately in our
*modern* desktop workloads, this quickly becomes ugly when we're dealing
with memory-limited systems. For example, a system with only 4GB RAM and a
64-bit OS, the need for swap quickly becomes apparent when dealing with
web browsers and other heavy workloads.

Now we calculate a safe swappiness, so that we actually make some use of it
in these more constrained scenarios, to prevent us getting in the nasty
situation of idle processes consuming all available memory until we're then
forcibly crunching the system with emergency page-outs.

This change doesn't yet write the new swappiness, but will in future.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>